renames action-config => actions-metrics
not sure action-config ever made sense, but this makes more sense now.
